fix a potential memory leak
This is very unlikely to occur in practice. Memory can be leaked
when TLS initialization fails when the client tries to connect
to the server. However, if this actually happens, it can happen
frequently and so accumulate to a large leak.

No report of such occurence from practice.

Detected by Coverity Scan, CID 266008.