bugfix: invalid memory allocation in offer generation
if memory allocation increase is requiered when building offers, the
new size was not properly reflected, leading to the guard variable
being used invalidly evaluated the next time and allocations redone.

This was very unlikely to cause serious problems. It was also
not expected that offers grow so much that the condition could be
triggerred at all, as potential offers are rather limited.

Detected by Coverity Scan, CID 266009