test framework: read correct receive pidfile
Make distcheck creates an own directory to process the tests
and the tests directory becomes read only during that process.
The receiver couldn't write a pidfile in the tests directory so
the test never could confirm that the receiver was running and
timed out.

The error messages are not redirected to librelp.out.log but written
to another file error.out.log, where it is easier to find them and
see what went wrong.

Closes https://github.com/rsyslog/librelp/issues/121