testbench: add realistic test with multiple messages sent
so far, only single messages were transmitted during testing. While
this is nice to check some basic features, it for sure is not sufficient
to detect many failure conditions.

This commit adds a first test which sends a larger number of messages
(50000 right now) and also extends the testbench framework with support
for this.