bugfix: do not send multiple open commands
multiple open commands can happen when a session was broken
and is being reestablished. Chances to experience this problem
increase with the number of sessions broken in a row. A trigger
is that the remote system accepts the TCP connection, but does
NOT send a response to the "open" command (aka "very short-lived
session).

Note that commit 12392015d4f8e99d6226ab827beb0b7836050efb tried
to "fix" this on the server side. Testing shows this is not possible.
That commit will be reverted soon.

This patch here fixes the root cause.

closes https://github.com/rsyslog/librelp/issues/129