bugfix: user pointer assigned too late, could lead to user segfault
When errors occurred early in object creation, the user pointer was not always
properly set and defaulted to NULL. If now an error was reported for some of
these errors, the NULL pointer was passed to the user application. If it
set a user pointer and tried to process it, it would segfault (or otherwise
malfunction).

The problem was not detected until recently, when we began to provide better
error messages.