CI: work around clang-9 compile error
The introduction of clang-9 compile testing revealed a number of
already-existing issues. This commit works around them, mostly
in a non-clean but acceptable way. Other patches shall follow.

We do this so that we can introduce clang-9 into the CI pipeline
and thus prevent additional new instances of the problems it
detects. While the workaround is not totally clean, it leaves
the code in a better shape than before.