AIX: in relpTcpRcv we need to set RETRY_recv if errno is 0
On AIX 10.x, when a recv call fails with len -1 and errno 0
we can still go into relpTCP_RETRY_recv mode. Testbench on
rsyslog side appears to be finally working with this fix.

On AIX 10.x, same problem as above appear to happen with
connect() as well not setting errno to EINPROGRESS on failure.

rsyslog ptcp driver does not check errno value at all, which is
why we did not have this problem on AIX with rsyslog.

Reduce test message count in receiver-abort.sh to 50000,
because otherwise the test can timeout on very slow machines.

Moved receive buffer in relpSessRcvData back to stack memory

closes: https://github.com/rsyslog/librelp/pull/239